Added simple chat functionality.
client (currently the grpc_cli tool) makes a chat request to the server, like so:
grpc_cli call localhost:8080 SendChat "message: 'sup'" --metadata
 "to:Mercury:from:Venus"

The metadata is used to associate the senders and recievers for the various chat operations.

The server has a hardcoded list of IP addresss that it uses to route the message to the appropriate client.

The client has a chat server that's listening for incoming chat requests.

Right now connections are one off -- they get created and destroyed every chat request.